requirements.txt
flask==3.0.2
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
airtable-python-wrapper==0.15.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
python-dateutil==2.8.2
flask-cors==4.0.0
pyairtable==2.3.0
pytz==2024.1 